% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_sbt.version.R, R/sbtmod.version.R
\name{.onAttach}
\alias{.onAttach}
\title{Function to return version number}
\usage{
.onAttach(libname, pkgname)

.onAttach(libname, pkgname)
}
\description{
Function to return version number

Function to return version number
}
