% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_get_all_files.R
\name{get_all_files}
\alias{get_all_files}
\title{Load _lab.rep files into R}
\usage{
get_all_files(directory)
}
\arguments{
\item{directory}{the directory that the _lab.rep files are in}
}
\value{
a list of objects
}
\description{
Goes to the specified directory and finds all the _lab.rep files. Loads them one by one into an R list of objects. This ensures that the reading in stage (which can take 3-10 minutes) need only be done once, and then subsequent function calls looping through an entire grid will run in a few seconds instead of minutes. Required the following files: something ending in "_lab.rep", assuming that the naming convention in the file has sections of outputs separated by $name comments so that it can be easily parsed by readList() function of Jon Schnute.
}
\examples{
data.example.sqrt <- get_all_files(directory = "arc/c1s1l1sqrt/")
}

