% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_get_MSY.R
\name{get_MSY_vals}
\alias{get_MSY_vals}
\title{Plot histograms of MSY values.}
\usage{
get_MSY_vals(data.objects, label)
}
\description{
Get the MSY results from each cell of the grid to see what the distribution of values is. Trevor A. Branch, started 5 September 2009, Busan, South Korea.
}

