% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_shaded.R
\name{panel_cor}
\alias{panel_cor}
\title{Panel correlations.}
\usage{
panel_cor(x, y, digits = 2, prefix = "", cex.cor)
}
\description{
Plot shaded plots. Put (absolute) correlations on the upper panels, with size proportional to the correlations.
}
\author{
Jim Ianelli.
}
