% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_shaded.R
\name{panel_smooth}
\alias{panel_smooth}
\title{Panel smooth.}
\usage{
panel_smooth(
  x,
  y,
  col = par("col"),
  bg = "green",
  pch = 24,
  cex = 1,
  col.smooth = "red",
  span = 1/3,
  iter = 3,
  ...
)
}
\description{
Panel smooth.
}
\author{
Jim Ianelli.
}
