% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_abundance_index_fits.R
\name{plot_aerial_survey_fit_comparison}
\alias{plot_aerial_survey_fit_comparison}
\title{Plot fits to aerial surveys}
\usage{
plot_aerial_survey_fit_comparison(
  data_objects,
  lev_files,
  scenario_names,
  year_label = NULL
)
}
\arguments{
\item{data_objects}{data objects}

\item{lev_files}{the lev files}

\item{scenario_names}{some names}
}
\value{
a ggplot object
}
\description{
Plots fits of the CCSBT model sbtmod22.tpl to the survey and index data: Aerial survey, troll survey, commercial spotter index, and CPUE index. Produces a four-panel plot with data and model fits to the data. 25 June 2009.
}
