% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_age_fits.R
\name{plot_Australian_age_fits}
\alias{plot_Australian_age_fits}
\title{Plot of model fit to the observed Australian age-frequencies by year.}
\usage{
plot_Australian_age_fits(labrep.file = "sbtmod_lab.rep", case_label = "")
}
\arguments{
\item{labrep.file}{name of the input file}

\item{case_label}{an additional label for the plot}
}
\description{
Fit of the model sbtmodxx.tpl to the data for southern bluefin tuna. Fits of the observed Australian age data to the predicted Australian age data.
}
\author{
Darcy Webber
}
