% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_age_fits.R
\name{plot_Australian_age_fits_old}
\alias{plot_Australian_age_fits_old}
\title{Plot of model fit to the observed Australian age-frequencies by year.}
\usage{
plot_Australian_age_fits_old(labrep.file = "sbtmod_lab.rep", case_label = "")
}
\arguments{
\item{labrep.file}{name of the input file}

\item{case_label}{an additional label for the plot}
}
\value{
a plot
}
\description{
Fit of the model sbtmodxx.tpl to the data for southern bluefin tuna. Fits of the observed Australian age data to the predicted Australian age data.
}
\examples{
pdf("figs/AussieAgeFts.pdf", width = 6, height = 11.5)
plot_aussie_age_fits(labrep.file = "sbtmod_lab.rep", case_label = "c1")
dev.off()
}
