% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_compare_plus_group.R
\name{plot_compare_plus_groups}
\alias{plot_compare_plus_groups}
\title{Plot plus groups.}
\usage{
plot_compare_plus_groups(compare.objects, ylim = c(0, 1), label)
}
\description{
Compare a variety of different model runs in terms of model numbers at age, Indonesian selectivity and natural mortality by age schedules.
}
