% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_harvest_rates.R
\name{plot_harvest_rates}
\alias{plot_harvest_rates}
\title{Boxplots harvest rates by age ranges}
\usage{
plot_harvest_rates(data.objects, lev.file = "base.lev", sample = TRUE,
  hline = 0.1)
}
\arguments{
\item{data.objects}{a list of the _lab.rep files from all grid cells}

\item{lev.file}{has the scenarios of 2000 sampled grid cells}

\item{sample}{sample}
}
\description{
Boxplots harvest rates by age ranges
}
\author{
Ana Parma, Darcy Webber
}
