% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_age_fits.R
\name{plot_Indonesian_age_fits}
\alias{plot_Indonesian_age_fits}
\title{Plot of model fit to the observed Indonesian age-frequencies by year.}
\usage{
plot_Indonesian_age_fits(data.object, case_label = "c1")
}
\value{
a plot
}
\description{
Fit of the model sbtmodxx.tpl to the data for southern bluefin tuna.
}

