% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_compare_plus_group.R
\name{plot_Indonesian_selectivity}
\alias{plot_Indonesian_selectivity}
\title{Plot Indonesian selectivity.}
\usage{
plot_Indonesian_selectivity(xx, year.list = c(1952, seq(1994, 2010, 2)),
  do.legend = F)
}
\description{
Plot Indonesian selectivity.
}

