% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Kobe.R
\name{plot_Kobe}
\alias{plot_Kobe}
\title{Plots U/Umsy vs B/Bmsy over time with confidence intervals.}
\usage{
plot_Kobe(MSY.stats, use.B10 = TRUE, catch.over.B = FALSE)
}
\description{
Umsy and Bmsy were calculated by msycalc.tpl using anual weights, selectivities and allocation
Ana Parma, Bali 2011
}
\author{
Ana Parma, Darcy Webber
}

