% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Kobe.R
\name{plot_Kobe_TRO}
\alias{plot_Kobe_TRO}
\title{Plots U/Umsy vs B/Bmsy over time with confidence intervals.}
\usage{
plot_Kobe_TRO(MSY.stats)
}
\arguments{
\item{MSY.stats}{MSY stats outputs}
}
\description{
Umsy and Bmsy were calculated by msycalc.tpl using anual weights, selectivities and allocation
Ana Parma, Bali 2011
}
\author{
Ana Parma, Darcy Webber
}
