% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_length_fits.R
\name{plot_length_fits_LL}
\alias{plot_length_fits_LL}
\title{Predicted length frequency for longline fishery 1 to the observed data.}
\usage{
plot_length_fits_LL(labrep.file = "sbtmod_lab.rep", case_label = "c1",
  LL = 1)
}
\description{
Fit of the model sbtmodxx.tpl to the data for southern bluefin tuna.
}
\examples{
pdf("figs/LengthsLL1.pdf", width = 8, height = 11.5)
LengthsLL1(labrep.file = "sbtmod_lab.rep", case_label = "c1")
dev.off()
}
