% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_length_fits.R
\name{plot_length_resid_LL}
\alias{plot_length_resid_LL}
\title{Length-frequency residuals}
\usage{
plot_length_resid_LL(labrep.file = "sbtmod_lab.rep", case_label = "c1", LL = 1)
}
\description{
Length-frequency residuals
}
