% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_NLL.R
\name{plot_NLL_by_M}
\alias{plot_NLL_by_M}
\title{Predicted length frequency for longline fishery 1 to the observed data.}
\usage{
plot_NLL_by_M(NLLtable, caption)
}
\description{
Plots the likelihoods for different likelihood components and the
total split out by different values of M10. Idea and original code:
Paige Eveson. Code modified and rewritten by Trevor A Branch 6
September 2009. Modified 24 June 2014 Trevor A Branch To split out
by different values of M10 instead of steepness.  Plot the
likelihoods by M10 values, using Paige's code and idea.
}
