% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_NLL.R
\name{plot_NLL_by_mM}
\alias{plot_NLL_by_mM}
\title{Plot the likelihoods by M1 values, using Paige Evesons code and idea. Plots the likelihoods for different likelihood components and the total split out by different values of M10.}
\usage{
plot_NLL_by_mM(NLLtable, caption)
}
\description{
Plot the likelihoods by M1 values, using Paige Evesons code and idea. Plots the likelihoods for different likelihood components and the total split out by different values of M10.
}
