% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_NLL.R
\name{plot_NLL_by_steepness}
\alias{plot_NLL_by_steepness}
\title{Plots the likelihoods for different likelihood components.}
\usage{
plot_NLL_by_steepness(NLLtable, caption)
}
\description{
Plot the likelihoods by steepness values, using Paige Evesons code and idea. Plots the likelihoods for different likelihood components and the total split out by different values of steepness. Code modified by Trevor A Branch 6 September 2009.
}

