% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_compare_plus_group.R
\name{plot_numbers_at_age}
\alias{plot_numbers_at_age}
\title{Plot numbers at age}
\usage{
plot_numbers_at_age(xx, min.age = 8, plot.years = c(1931, 2008))
}
\description{
Plot of numbers at age.
}

