% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_PEN.R
\name{plot_pen_by_steepness}
\alias{plot_pen_by_steepness}
\title{Plot the likelihoods by steepness values, using Paige's code and idea.}
\usage{
plot_pen_by_steepness(pentable, caption)
}
\description{
Plots the penalties for different penalty components and the total
split out by different values of steepness. 
Idea and original code: Paige Eveson
Code modified and rewritten by Trevor A Branch 6 September 2009  tbranch@gmail.com
Code blatantly copied and minimally adapted by Rich Hillary Jul 2012
read in all the files into an object (takes time)
source("get.all.data.r")
data.M4constrained <- get.all.files(directory="M4constrained\\arc\\c1s1l1sqrt")
get table of likelihood components and steepness values (quick)
source("TableLikelihoodComponents.r")
LikeM4const  <- likelihood.table(data.objects = data.M4constrained)
LikeM4unconst <- likelihood.table(data.objects = data.M4unconst)
LikeMpower <- likelihood.table(data.objects = data.Mpower)
}

