% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot of recruitment}
\usage{
plot_recruitment(data.objects, lev.file = "base.lev", scenario_name,
  sample = TRUE)
}
\arguments{
\item{data.objects}{a list of the _lab.rep files from all grid cells}

\item{lev.file}{has the scenarios of 2000 sampled grid cells}

\item{scenario_name}{a vector of name labels for the scenarios being plotted}
}
\description{
Plot of recruitment
}

