% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_SR.R
\name{plot_recruitment_and_ssb}
\alias{plot_recruitment_and_ssb}
\title{Plot recruitment and SSB}
\usage{
plot_recruitment_and_ssb(data.objects)
}
\arguments{
\item{data.objects}{a list of the _lab.rep files from all grid cells}
}
\description{
Plot recruitment and SSB
}
\author{
Ana Parma, Darcy Webber
}
