% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_SR.R
\name{plot_SR}
\alias{plot_SR}
\title{Plot SR}
\usage{
plot_SR(data.objects)
}
\arguments{
\item{data.objects}{a list of the _lab.rep files from all grid cells}
}
\description{
Plot SR
}
\author{
Ana Parma, Darcy Webber
}
