% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_compare_plus_group.R, R/om_plot_SSB.R
\name{plot_SSB}
\alias{plot_SSB}
\title{Plot spawning stock biomass.}
\usage{
plot_SSB(data.objects, lev.file = "base.lev", sample = TRUE)

plot_SSB(data.objects, lev.file = "base.lev", sample = TRUE)
}
\arguments{
\item{data.objects}{a list of the _lab.rep files from all grid cells}

\item{lev.file}{has the scenarios of 2000 sampled grid cells}
}
\description{
Plot of spawning stock biomass (SSB).
}
