% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_SSB.R
\name{plot_SSB_compare}
\alias{plot_SSB_compare}
\title{SSB by year comparison}
\usage{
plot_SSB_compare(data_objects, lev_files, scenario_names, sample = TRUE)
}
\arguments{
\item{data.objects}{a list of the _lab.rep files from all grid cells}

\item{lev.file}{has the scenarios of 2000 sampled grid cells}
}
\description{
SSB by year comparison
}
