% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_tagging_fits.R
\name{plot_tagging_fits}
\alias{plot_tagging_fits}
\title{Plots the fits to the tagging data}
\usage{
plot_tagging_fits(data.object, case_label = "c1")
}
\description{
Plots the fits to the tagging data for sbtmod22.tpl outputs from a _lab.rep file, assuming that the naming convention in the file has sections of outputs separated by $name comments so that it can be easily parsed by readList() function of Jon Schnute. RUN this by typing , expects subdirectory figs/
}
\examples{
pdf("figs/Tagfits combine taggers.pdf", width = 6, height = 8)
plot_tagging_fits(data.object = x[[1]], case_label = "example")
dev.off()
}

