% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_tagging_fits.R
\name{plot_tagging_fits_pooled}
\alias{plot_tagging_fits_pooled}
\title{Plots the fits to the pooled tagging data.}
\usage{
plot_tagging_fits_pooled(data.object, ages = 2:8, years = 1992:1997,
  case_label = "c1s1l1orig.5_h1m1M1O1C2a1")
}
\description{
Plots the fits to the pooled tagging data for sbtmod22.tpl outputs from a _lab.rep file.
}
\examples{
pdf("figs/Tagfits pooled.pdf", width = 3, height = 8)
# Does not work on sbtmod21 (since the tagging likelihood there is on pooled taggging data)
tagging.fits.pooled(data.object = x[[1]], ages = 2:8, years = 1992:1997, case_label="c1")
dev.off()
}

