% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_tagging_fits.R
\name{plot_tagging_fits_pooled_compare}
\alias{plot_tagging_fits_pooled_compare}
\title{Plots side by side fits to tagging data from two lab_rep files.}
\usage{
plot_tagging_fits_pooled_compare(compare.objects, case.label, ages = 2:8,
  years = 1992:1997)
}
\description{
Assumes that the _lab.rep files have been read in using code like x <- readList(file="example_lab.rep") and that compare.objects is a list.
}
