% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_tagging_fits.R
\name{plot_tagging_fits_subsequent}
\alias{plot_tagging_fits_subsequent}
\title{Plots the fits to the tagging data for sbtmod22.tpl}
\usage{
plot_tagging_fits_subsequent(
  data.object,
  case_label = "c1s1l1orig.5_h1m1M1O1C2a1"
)
}
\description{
outputs from a _lab.rep file, assuming that the naming convention in the file has sections of 
outputs separated by $name comments so that it can be easily parsed by readList() function of Jon Schnute 
RUN this by typing source("Tagging fits subsequent.r"), expects subdirectory figs
}
