% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_abundance_index_fits.R
\name{plot_troll_survey_fit}
\alias{plot_troll_survey_fit}
\title{Plot fits to troll surveys}
\usage{
plot_troll_survey_fit(data.object, lab.cex = 1.1)
}
\description{
Plots fits of the CCSBT model sbtmod22.tpl to the survey and index data: Aerial survey, troll survey, commercial spotter index, and CPUE index. Produces a four-panel plot with data and model fits to the data. 25 June 2009.
}

