% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_plot_abundance_index_fits.R
\name{plot_troll_survey_fit_comparison}
\alias{plot_troll_survey_fit_comparison}
\title{Plot fits to troll surveys}
\usage{
plot_troll_survey_fit_comparison(
  data_objects,
  lev_files,
  scenario_names,
  year_label = NULL
)
}
\arguments{
\item{data_objects}{data objects}

\item{lev_files}{the lev files}

\item{scenario_names}{some names}
}
\value{
a ggplot object
}
\description{
Plot fits to troll surveys
}
