% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_table_likelihood.R
\name{table_likelihood}
\alias{table_likelihood}
\title{Table of the likelihood components for a specified set of _lab.rep files}
\usage{
table_likelihood(data.objects)
}
\description{
Table of the likelihood components for a specified set of _lab.rep files
}
